requirements.txt
appdirs==1.4.4
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
distlib==0.3.1
Django==3.0
django-allauth==0.40.0
django-cors-headers==3.7.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-framework-roles==0.6
djangorestframework==3.10.3
drf-yasg==1.20.0
ecdsa==0.15
filelock==3.0.12
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.20.2
oauthlib==3.1.0
packaging==20.8
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-google-places==1.4.2
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
pyzmq==20.0.0
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==20.4.4
virtualenv-clone==0.5.4
